added coments to service classes added OrderFirebaseService and DeviceUtils
